# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OGM Gönüllü Yönetim Sistemi API",
    "version": "0.1.0",
    "description": "T.C. Orman Genel Müdürlüğü Gönüllü Yönetim Sistemi backend API'si. Auth (e-Devlet + Telefon OTP), onboarding, profil, ana sayfa, yangın bildirimi ve referans veri uçlarını içerir.",
    "contact": {
      "name": "OGM Gönüllü Backend",
      "email": "gonullu-sistem@ogm.gov.tr"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com/v1",
      "description": "API v1"
    },
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com",
      "description": "Kök (health vb.)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Erişim token'ı (`Authorization: Bearer <token>`)"
      },
      "registrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Telefon doğrulama sonrası kayıt formu için kısa süreli token"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Panel admin anahtarı (ADMIN_API_KEY)"
      },
      "officerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Saha amiri anahtarı (OFFICER_API_KEY)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "validation_error"
              },
              "message": {
                "type": "string",
                "example": "Form alanlarında hata var"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "ReferenceItem": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "example": "A+"
          },
          "label": {
            "type": "string",
            "example": "A+"
          }
        }
      },
...
      "ActiveMissionSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "example": "LOJİSTİK DESTEK"
          },
          "title": {
            "type": "string"
          },
          "shortLocation": {
            "type": "string",
            "example": "Muğla / Marmaris"
          },
          "iconName": {
            "type": "string",
            "enum": [
              "water",
              "helmet",
              "tool",
              "first-aid"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "staffed"
            ]
          },
          "userStatus": {
            "type": "string",
            "enum": [
              "not_joined",
              "accepted",
              "on_site"
            ]
          }
        }
      },
      "ActiveMissionDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ActiveMissionSummary"
          },
          {
            "type": "object",
            "properties": {
              "regionLabel": {
                "type": "string"
              },
              "fullTitle": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "gallery": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "needs": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "stats": {
                "type": "object",
                "properties": {
                  "volunteers": {
                    "type": "integer"
                  },
                  "hectares": {
                    "type": "number"
                  }
                }
              },
              "locationLabel": {
                "type": "string"
              },
              "startedAt": {
                "type": "string",
                "format": "date-time"
              },
              "coordinates": {
                "type": "object",
                "properties": {
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  }
                }
              },
              "coverageRadiusKm": {
                "type": "number"
              },
              "operational": {
                "type": "object",
                "properties": {
                  "meetingPoint": {
                    "type": "string"
                  },
                  "requiredEquipment": {
                    "type": "string"
                  }
                }
              },
              "announcements": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "publishedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "severity": {
                      "type": "string",
                      "enum": [
                        "info",
                        "alert"
                      ]
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "OnlineTraining": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "durationMin": {
            "type": "integer"
          },
          "iconTone": {
            "type": "string",
            "enum": [
              "primary",
              "tertiary"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "not_started",
              "in_progress",
              "completed"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
      "SahaTraining": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "17:00"
          },
          "instructorName": {
            "type": "string"
          },
          "instructorAvatar": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "cover": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "availableSeats": {
            "type": "integer"
          },
          "seatStatus": {
            "type": "string",
            "enum": [
              "available",
              "last_seats"
            ]
          },
          "applied": {
            "type": "boolean"
          }
        }
      },
...
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "taskCalls": {
            "type": "boolean"
          },
          "trainings": {
            "type": "boolean"
          },
          "announcements": {
            "type": "boolean"
          },
          "distance": {
            "type": "object",
            "properties": {
              "km": {
                "type": "integer"
              },
              "min": {
                "type": "integer"
              },
              "max": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation hatası",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Kaynak bulunamadı",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "security": [],
  "paths": {
...
    "/missions/active": {
      "get": {
        "tags": [
          "Missions"
        ],
        "summary": "Aktif görev listesi (Anasayfa)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aktif görevler",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveMissionSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/missions/active/{id}": {
      "get": {
        "tags": [
          "Missions"
        ],
        "summary": "Aktif görev detayı",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Görev detayı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveMissionDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/missions/active/{id}/join": {
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "Göreve katıl (userStatus → accepted)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Katılım onaylandı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "userStatus": {
                      "type": "string",
                      "example": "accepted"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "equipment_required"
          },
          "409": {
            "description": "mission_full | already_joined"
          }
        }
      }
    },
    "/missions/active/{id}/scan": {
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "Saha amiri QR tarama (userStatus → on_site)",
        "description": "Mobil uygulamadan ÇAĞRILMAZ. Saha amiri (officer) `x-api-key` veya\n`role=officer` token ile çağırır. Opsiyonel `token` HMAC imzasıdır (B.1).\n",
        "security": [
          {
            "officerApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "scannedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "token": {
                    "type": "string",
                    "description": "opsiyonel HMAC imzası"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "on_site'a geçildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "userStatus": {
                      "type": "string",
                      "example": "on_site"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/missions/active/{id}/photos": {
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "Görev fotoğrafı yükle (yalnızca on_site)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission alındı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "submissionId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    },
                    "submittedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "not_on_site"
          }
        }
      }
    },
...
    "/users/me/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Bildirim tercihleri",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tercihler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Bildirim tercihlerini güncelle",
        "description": "Request `distanceKm` (flat) gönderilir; response `distance:{km,min,max}` (nested) döner.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskCalls": {
                    "type": "boolean"
                  },
                  "trainings": {
                    "type": "boolean"
                  },
                  "announcements": {
                    "type": "boolean"
                  },
                  "distanceKm": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Güncellenmiş tercihler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/reference/{category}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Kategori için referans listesini döndürür",
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "kan-grubu",
                "ogrenim",
                "meslek",
                "hobiler",
                "yakinlik",
                "ulkeler"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReferenceItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/trainings/online": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "summary": "Online eğitim listesi (kullanıcı bazlı progress)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Online eğitimler",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnlineTraining"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trainings/saha": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "summary": "Saha eğitim listesi",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Saha eğitimleri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SahaTraining"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trainings/{id}/applications": {
      "post": {
        "tags": [
          "Trainings"
        ],
        "summary": "Saha eğitimine başvur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Başvuru alındı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "applicationId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "already_applied"
          },
          "410": {
            "description": "training_full | training_closed"
          }
        }
      }
    },
...
  }
}